RetailPriceDiscount activity

+++ src/Kooboo.Commerce.Web/Areas/Kooboo.Commerce.Promotions.Policies.Default/DefaultPromotionPolicyData.cs
@@ -14,7 +14,7 @@ namespace Kooboo.Commerce.Promotions.Policies.Default
 
         public decimal DiscountAmount { get; set; }
 
-        public float DiscountPercent { get; set; }
+        public int DiscountPercent { get; set; }
 
         public string Serialize()
         {
